+++ database/migrations/2020_08_02_115431_rename_beneficiary_to_attendee_in_money_transactions.php
@@ -1,0 +1,32 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+class RenameBeneficiaryToAttendeeInMoneyTransactions extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('money_transactions', function (Blueprint $table) {
+            $table->renameColumn('beneficiary', 'attendee');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('money_transactions', function (Blueprint $table) {
+            $table->renameColumn('attendee', 'beneficiary');
+        });
+    }
+}